requirements.txt
# Web Framework
Flask==3.0.3
Werkzeug==3.0.1
gunicorn==23.0.0

# Database
pymongo==4.10.1
python-dotenv==1.0.1  

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Data Visualization
plotly==5.24.1
matplotlib==3.8.3
seaborn==0.13.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
urllib3==2.2.3

# Utilities
python-dateutil==2.9.0
typing-extensions==4.12.2